Update de token

+++ resources/Item.py
@@ -1,5 +1,6 @@
 import sqlite3
 from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from flask_restful import Resource, reqparse
 
 from models.ItemModel import ItemModel
@@ -17,7 +18,7 @@ class Item(Resource):
             return row.json()
         return {"message": "Item not found"}
 
-    @jwt_required()
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": "Item already exists"}, 400
@@ -32,8 +33,12 @@ class Item(Resource):
 
         return {"message": "Item Created successfully"}
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -53,5 +58,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return [item.json() for item in ItemModel.query.all()]
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return items
+        return {
+            'message': 'More data if login',
+            'items': [item['name'] for item in items]
+        }
